backend/pyproject.toml
[project]
name = "hiremesh-backend"
version = "0.1.0"
description = "Hiremesh API and workers"
requires-python = ">=3.13"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg[binary]>=3.2",
    "pydantic[email]>=2.9",
    "pydantic-settings>=2.6",
    "passlib[argon2]>=1.7.4",
    "argon2-cffi>=23.1",
    "pyjwt>=2.9",
    "python-multipart>=0.0.12",
    "celery[redis]>=5.4",
    "redis>=5.2",
    "boto3>=1.35",
    "litellm>=1.50",
    "pypdf>=5.0",
    "python-docx>=1.1",
    "pgvector>=0.3",
    "numpy>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.7",
    "mypy>=1.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]
